// sdk/src/generated/types/oracleSource.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  fixDecoderSize,
  fixEncoderSize,
  getBytesDecoder,
  getBytesEncoder,
  getDiscriminatedUnionDecoder,
  getDiscriminatedUnionEncoder,
  getStructDecoder,
  getStructEncoder,
  getU16Decoder,
  getU16Encoder,
  getU32Decoder,
  getU32Encoder,
  type Codec,
  type Decoder,
  type Encoder,
  type GetDiscriminatedUnionVariant,
  type GetDiscriminatedUnionVariantContent,
  type ReadonlyUint8Array,
} from "@solana/kit";

export type OracleSource =
  | { __kind: "Scope"; index: number }
  | { __kind: "Pyth"; feedId: ReadonlyUint8Array; maxAge: number };

export type OracleSourceArgs = OracleSource;

export function getOracleSourceEncoder(): Encoder<OracleSourceArgs> {
  return getDiscriminatedUnionEncoder([
    ["Scope", getStructEncoder([["index", getU16Encoder()]])],
    [
      "Pyth",
      getStructEncoder([
        ["feedId", fixEncoderSize(getBytesEncoder(), 32)],
        ["maxAge", getU32Encoder()],
      ]),
    ],
  ]);
}

export function getOracleSourceDecoder(): Decoder<OracleSource> {
  return getDiscriminatedUnionDecoder([
    ["Scope", getStructDecoder([["index", getU16Decoder()]])],
    [
      "Pyth",
      getStructDecoder([
        ["feedId", fixDecoderSize(getBytesDecoder(), 32)],
        ["maxAge", getU32Decoder()],
      ]),
    ],
  ]);
}

export function getOracleSourceCodec(): Codec<OracleSourceArgs, OracleSource> {
  return combineCodec(getOracleSourceEncoder(), getOracleSourceDecoder());
}

// Data Enum Helpers.
export function oracleSource(
  kind: "Scope",
  data: GetDiscriminatedUnionVariantContent<
    OracleSourceArgs,
    "__kind",
    "Scope"
  >,
): GetDiscriminatedUnionVariant<OracleSourceArgs, "__kind", "Scope">;
export function oracleSource(
  kind: "Pyth",
  data: GetDiscriminatedUnionVariantContent<OracleSourceArgs, "__kind", "Pyth">,
): GetDiscriminatedUnionVariant<OracleSourceArgs, "__kind", "Pyth">;
export function oracleSource<K extends OracleSourceArgs["__kind"], Data>(
  kind: K,
  data?: Data,
) {
  return Array.isArray(data)
    ? { __kind: kind, fields: data }
    : { __kind: kind, ...(data ?? {}) };
}

export function isOracleSource<K extends OracleSource["__kind"]>(
  kind: K,
  value: OracleSource,
): value is OracleSource & { __kind: K } {
  return value.__kind === kind;
}